+++ lib/database_consistency/checkers/presence_validation_checker.rb
@@ -1,11 +1,13 @@
 module DatabaseConsistency
-  module Comparators
-    # The comparator class for {{ActiveModel::Validations::PresenceValidator}}
-    class PresenceComparator < BaseComparator
+  module Checkers
+    # This class checks PresenceValidator
+    class PresenceValidationChecker < BaseChecker
       WEAK_OPTIONS = %i[allow_nil allow_blank if unless].freeze
       # Message templates
       CONSTRAINT_MISSING = 'should be required in the database'.freeze
       POSSIBLE_NULL = 'is required but possible null value insert'.freeze
+
+      private
 
       # Table of possible statuses
       # | allow_nil/allow_blank/if/unless | database | status |
@@ -14,23 +16,38 @@ module DatabaseConsistency
       # | at least one provided           | optional | ok     |
       # | all missed                      | required | ok     |
       # | all missed                      | optional | fail   |
-      def compare
+      #
+      # We skip check when:
+      #  - there is no column in the database with given name
+      def check
+        return unless column
+
         can_be_null = column.null
         has_weak_option = validator.options.slice(*WEAK_OPTIONS).any?
 
         if can_be_null == has_weak_option
-          result(:ok, message)
+          report_template(:ok)
         elsif can_be_null
-          result(:fail, message(CONSTRAINT_MISSING))
+          report_template(:fail, CONSTRAINT_MISSING)
         else
-          result(:fail, message(POSSIBLE_NULL))
+          report_template(:fail, POSSIBLE_NULL)
         end
       end
 
-      private
+      def column
+        @column ||= Helper.find_field(table_or_model, column_or_attribute.to_s)
+      end
 
-      def message(template = nil)
-        Helper.message(model, column, template)
+      def column_or_attribute_name
+        column_or_attribute.to_s
+      end
+
+      def table_or_model_name
+        table_or_model.name.to_s
+      end
+
+      def validator
+        opts[:validator]
       end
     end
   end