+++ spec/spec_helper.rb
@@ -1,6 +1,8 @@
 require 'bundler/setup'
 require 'database_consistency'
 
+require 'database_context'
+
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = '.rspec_status'
@@ -10,5 +12,9 @@ RSpec.configure do |config|
 
   config.expect_with :rspec do |c|
     c.syntax = :expect
+  end
+
+  def file_fixture(path)
+    File.join('spec/fixtures/files/', path)
   end
 end